Almost Headset Ready
Since triggers aren't working I have resorted to boxcasting as triggers. This is only done on fixed update as if it were physics, and also only when the object is being manipulated & on correct step (might remove correct step only depending on bosses). Once an action is performed for long enough the manager is advanced to the next step. Added particle/sound (wip) effects as indication. Working on other systems of triggers tomorrow and then building to headset.